tease out Laplace into new file laplace.py
