+++ edward/inferences/map.py
@@ -7,7 +7,7 @@ import tensorflow as tf
 
 from edward.inferences.variational_inference import VariationalInference
 from edward.models import RandomVariable, PointMass
-from edward.util import copy, hessian
+from edward.util import copy
 
 
 class MAP(VariationalInference):
@@ -143,28 +143,3 @@ class MAP(VariationalInference):
     grads = tf.gradients(loss, [v._ref() for v in var_list])
     grads_and_vars = list(zip(grads, var_list))
     return loss, grads_and_vars
-
-
-class Laplace(MAP):
-  """Laplace approximation.
-
-  It approximates the posterior distribution using a normal
-  distribution centered at the mode of the posterior.
-  """
-  def __init__(self, *args, **kwargs):
-    super(Laplace, self).__init__(*args, **kwargs)
-
-  def finalize(self):
-    """Function to call after convergence.
-
-    Computes the Hessian at the mode.
-    """
-    # use only a batch of data to estimate hessian
-    x = self.data
-    z = {z: qz.value() for z, qz in six.iteritems(self.latent_vars)}
-    var_list = tf.get_collection(tf.GraphKeys.TRAINABLE_VARIABLES,
-                                 scope='posterior')
-    inv_cov = hessian(self.model_wrapper.log_prob(x, z), var_list)
-    print("Precision matrix:")
-    print(inv_cov.eval())
-    super(Laplace, self).finalize()
